requirements.txt
# GC-ATLAS data pipeline dependencies
# ------------------------------------------------------------
# Install with:  pip install -r requirements.txt
# ------------------------------------------------------------

# Copernicus CDS client
cdsapi>=0.7

# Scientific / array stack
numpy>=1.24,<2.0
scipy>=1.12
xarray>=2024.1
dask[array]>=2024.2

# NetCDF / Zarr / GCS I/O
netCDF4>=1.6
h5netcdf>=1.3
h5py>=3.10
zarr>=2.16,<3.0
fsspec>=2024.2
gcsfs>=2024.2

# Config + CLI niceties
pyyaml>=6.0
tqdm>=4.66

# Backend API (Phase 2+)
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
gunicorn>=21.2
